Updated install procedure
In an effort to make this easier to install/repair, I have changed the procedure of the update-channels-from-git.sh file.  

This now takes on the form of an "installer" which will move any files that are currently on GitHub to their corrsponding sections (main-dir, channel-dir, or both-dir). NOTE: this will NOT overwrite files you have made on your own to customize, and all databases, xml files, config files, and tokens will be preserved.