manual.sh minor bug fix
Issues were occurring where manual.sh could not accept incorrect channel input.  Now, if an invalid channel is input, you will simply be diverted to your first channel.  This can be changed later in many ways, one of which is to simply kill the script; I did not choose this since going to channel 1 kind of serves as an "error report" since you didn't make it to the channel you requested.