Global_DatabaseUpdate Quick Fix
At the request of some, I have modified the Global_DatabaseUpdate file to import the previous database's daily_schedule.  

NOTE: If you end up experiencing issues with your shows playing, this may be because your new database does not have the resources to play the listed shows.  May be time to run -g at that point.