Update of Schedule Controller
Bug reports were coming in where like-named shows were not all being played.

This fix attempts to solve the problem, by checking show title AND duration.  This ensures that you find exactly the right thing to play.

The fix has also been translated to the website creation as well, so the correct artwork will be portrayed for each movie/show.