Commercial logic improvement; website creation changes
Modifications were made to allow for commercials to be generated for the last show before midnight. This had been an issue in the past, and should be fixed with this update.

Further modifications were also made to the -m flag script, hopefully removing the issues surrounding multiple items being declared running.