Inital major Python3 change
In an effort to update this code to fit standards of Python3, changes have been made to code allowing it to run.

NOTE: At this point the basics have been tested.  i.e. -g -xml -u -m -r in that order.  More extensive testing should occur before we throw this to master.