Added "secondary" time-shift mode to allow pre-empting by long-running shows
This mode can be used by using the value "secondary" for time-shift in the schedule xml file. This will maintain a show or movie's start time, but allow the previous item to complete before starting the scheduled item "already in progress." If the previous item runs over the running time of the 'secondary' scheduled show or movie, it won't be scheduled for that day (may use overlap-max values in the future to allow manual overlap cutoff.