Added "Playlist as TV Show" Functionality
Scan your database with -u and it will include your plex playlists. Call these the same way you would a TV show in the schedule XML, using the title of the Playlist in place of the title of the TV show. It will cycle through the playlist once per call, in the same manner as a show in your schedule.