Added function to manually set last episodes
Using the -le argument, a show can be chosen to manually update the "last episode played" value in the shows database. I added this due to plex sometimes changing a show title, causing a loss in last episode data on a pseudo channel database rescan (for example, it changed Doctor Who (2005) to just Doctor Who). This function will prompt the user for the episode name (as it appears in Plex) and a season and episode number. It will set that episode as the "last played" so pseudo channel will schedule the episode after on the next daily update. 

Note: If Plex has updated a show title to be different than what's in the database, you will need to manually update the schedule to account for the change.