Improved Scheduling
Added functionality that allows commercials to be scheduled up to reset time (instead of up to the last scheduled program) and fixed the bug that prevented commercials from being scheduled between the last program before midnight and midnight. Use of this functionality will depend on the reset time in your pseudo_config file being accurate to your cron settings.