Added input option to last episode selection
In the --last-episode/-le function to manually set the last scheduled episode in the database for any show or playlist, the ability to use the unique plex id (ratingKey in the API, or the unique digits near the end (before the Plex Token) of the URL when going from File Info -> View XML Data on an episode page within Plex) instead of season and episode numbers. This better allows support for this function with Playlists, which may have multiple files with the same season and episode numbers.